feat: Re-add Debugify (using Fabric dep. override)
Since Debugify allows for a smoother experience, I'd like to keep it in
the pack. The 1.20 version is bugged, but you can force the patched
(1.20.1) version to run on whatever Minecraft version you want!

For more information, see the [Fabric documentation on the
subject](https://fabricmc.net/wiki/tutorial:dependency_overrides)!